RDKCOM-3134 RDKBCMF-290 Utopia 64bit build failures
Build errors in RPi 64bit builds in the utopia component:
-Werror=int-to-pointer-cast
-Werror=pointer-to-int-cast
-Werror=format
-Werror=format-truncation
-Werror=incompatible-pointer-types
-Werror=int-conversion

Reason for change: RDKBCMF-290 Utopia 64bit build failures
Test Procedure: None.
Risks: Low.

(cherry picked from commit 48cfd0a501e4d410298c503f25539f591c00098c)

+++ source/service_multinet/service_multinet_swfab_ep.c
@@ -132,9 +132,9 @@ int ep_get_entity_vid_portMembers(int vid, int entity, char* memberPortNames[], 
     
     while (token) {
 			
-	if((int)numPorts == MAX_ADD_PORTS )
+	if((long int)numPorts == MAX_ADD_PORTS )
 	{
-		MNET_DEBUG("ep_get_entity_vid_portMembers  numPorts is exceeded [%d]\n"COMMA (int)numPorts)
+		MNET_DEBUG("ep_get_entity_vid_portMembers  numPorts is exceeded [%ld]\n"COMMA (long int)numPorts)
 		break;
 	}
         memberPortNames[*numPorts] = buf + offset;
@@ -210,9 +210,9 @@ int ep_get_trunkPort_vidMembers(int vid, char* portNames[], int* numPorts, char 
     *numPorts = 0;
     while (token) {
 			
-	if((int)numPorts == MAX_ADD_PORTS )
+	if((long int)numPorts == MAX_ADD_PORTS )
 	{
-		MNET_DEBUG("ep_get_entity_vid_portMembers numPorts is exceeded[%d]\n"COMMA (int)numPorts)
+		MNET_DEBUG("ep_get_entity_vid_portMembers numPorts is exceeded[%ld]\n"COMMA (long int)numPorts)
 		break;
 	}
         portNames[*numPorts] = buf + offset;
